modularising the parseUploaded docs and ccdDocumentCollection functions
