+++ src/main/steps/respond-to-claim/upload-document/index.ts
@@ -6,59 +6,12 @@ import type { PossessionClaimResponse } from '@interfaces/ccdCaseData.model';
 import type { CdamDocument } from '@interfaces/documentUpload.interface';
 import type { StepDefinition } from '@interfaces/stepFormData.interface';
 import { createFormStep, getTranslationFunction } from '@modules/steps';
-import { ACCEPT_ATTRIBUTE_EXTENSIONS, UPLOAD_MAX_FILE_SIZE_MB } from '@utils/documentUploadValidation';
-
-function parseUploadedDocuments(body: Record<string, unknown>): CdamDocument[] {
-  const raw = body['uploadedDocuments[]'];
-  if (!raw) {
-    return [];
-  }
-
-  const items = Array.isArray(raw) ? raw : [raw];
-  const documents: CdamDocument[] = [];
-
-  for (const item of items) {
-    try {
-      const parsed = typeof item === 'string' ? JSON.parse(item) : item;
-      if (parsed?.document_url && parsed?.document_binary_url && parsed?.document_filename) {
-        const doc: CdamDocument = {
-          document_url: parsed.document_url,
-          document_binary_url: parsed.document_binary_url,
-          document_filename: parsed.document_filename,
-        };
-        if (typeof parsed.content_type === 'string' && parsed.content_type) {
-          doc.content_type = parsed.content_type;
-        }
-        const sizeNum = typeof parsed.size === 'number' ? parsed.size : Number(parsed.size);
-        if (!Number.isNaN(sizeNum) && sizeNum >= 0) {
-          doc.size = sizeNum;
-        }
-        documents.push(doc);
-      }
-    } catch {
-      // Skip malformed entries
-    }
-  }
-
-  return documents;
-}
-
-function toCcdDocumentCollection(docs: CdamDocument[]): CcdCollectionItem<CcdDocumentReference>[] {
-  return docs.map(doc => {
-    const value: CcdDocumentReference = {
-      document_url: doc.document_url,
-      document_binary_url: doc.document_binary_url,
-      document_filename: doc.document_filename,
-    };
-    if (doc.content_type !== undefined && doc.content_type !== null) {
-      value.content_type = doc.content_type;
-    }
-    if (doc.size !== undefined && doc.size !== null) {
-      value.size = doc.size;
-    }
-    return { value };
-  });
-}
+import {
+  ACCEPT_ATTRIBUTE_EXTENSIONS,
+  UPLOAD_MAX_FILE_SIZE_MB,
+  parseUploadedDocuments,
+  toCcdDocumentCollection,
+} from '@utils/documentUploadValidation';
 
 export const step: StepDefinition = createFormStep({
   stepName: 'upload-document',
